Using floating point numbers

+++ CPractice/HelloFromC/Main.c
@@ -2,7 +2,7 @@
 
 void main()
 {
-	int fahr, celcius;
+	float fahr, celcius;
 	int lower, upper, step;
 
 	lower = 0;
@@ -11,8 +11,8 @@ void main()
 
 	fahr = lower;
 	while (fahr <= upper) {
-		celcius = 5 * (fahr - 32) / 9;
-		printf("%3d\t%6d\n", fahr, celcius);
+		celcius = 5.0 * (fahr - 32.0) / 9.0;
+		printf("%3.0f\t%6.1f\n", fahr, celcius);
 		fahr = fahr + step;
 	}
